refactor(agent): extract buildConnectionConfig to shared utility
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
